# Adding dummyHash to DirectGrant request in case user does not exists.…