… Fix dummyHash for normal login requests

closes #12298

### Summary of changes

- PR adds dummyHash to direct grant requests
- PR fixes dummyHash for regular login requests, which was broken due the Argon2 change

### Details

Before this PR, it is easily possible to enumerate Keycloak users as there is significant difference between the DirectGrant (OAuth2 resource owner password credentials) request of:

- Existing username, but invalid password
- Non-existing username

The latter is much faster (around 2-3 times) and hence can be easily used to detect if user exists.

This PR tries to mitigate the difference to simulate the password validation for non-existing user as well (by using `dummyHash` method). This method is already used for regular browser user logins.

However during the work on this, I figured that `AbstractUsernameFormAuthenticator.dummyHash` method was broken for the Argon2 provider because it used method `PasswordHashProvider.encode(String rawPassword, int iterations)`, which in case of `Argon2PasswordHashProvider` does nothing (it just return the given password). I've fixed this to use `PasswordHashProvider.encodedCredential` method (which works as expected for both Argon2 and Pbkdf2 providers).

Also I've made the method `PasswordHashProvider.encode` deprecated to avoid similar issues in the future (it was used anyway just in the test and in this `dummyHash` method).

I am adding the numbers on my laptop (PR adds the difference making it harder to reliably enumerate users). There is however still one remaining issue that very first request is slower for the existing user as long as the user is NOT in the cache as it seems there are multiple DB lookups (like the request for retrieve roles, password credentials etc). I did not addressed this one, but rather consider it as a follow-up as AFAIK there is some effort in progress to optimize lookup the user to avoid using multiple queries.

```
BEFORE CHANGE:

10:09:14,445 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] Invalid password 1: 154 ms
10:09:14,531 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] Invalid password 2: 86 ms
10:09:14,631 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] Invalid password 3: 100 ms
10:09:14,712 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] Invalid password 4: 81 ms
10:09:14,790 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] Invalid password 5: 78 ms
10:09:14,828 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] User not found 1: 38 ms
10:09:14,867 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] User not found 2: 39 ms
10:09:14,907 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] User not found 3: 40 ms
10:09:14,949 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] User not found 4: 42 ms
10:09:14,989 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] User not found 5: 39 ms
10:09:14,991 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] calling all TestCleanup
java.lang.AssertionError: Times in ms of 5 attempts: For invalid password: 499. For invalid username: 198


AFTER CHANGE:

12:05:49,383 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] Invalid password 1: 150 ms
12:05:49,466 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] Invalid password 2: 82 ms
12:05:49,545 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] Invalid password 3: 79 ms
12:05:49,639 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] Invalid password 4: 94 ms
12:05:49,745 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] Invalid password 5: 105 ms
12:05:49,839 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] User not found 1: 94 ms
12:05:49,914 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] User not found 2: 75 ms
12:05:49,978 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] User not found 3: 64 ms
12:05:50,048 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] User not found 4: 92 ms
12:05:50,124 INFO  [org.keycloak.testsuite.oauth.ResourceOwnerPasswordCredentialsGrantTest] User not found 5: 76 ms
```