@jonkoops 2024-05-13
@mposolda Have you given any thought to https://github.com/keycloak/keycloak/pull/29304#pullrequestreview-2052387943?