@ahus1 2024-05-13
> I am wondering, should a per-user global lock be implemented so that an attacker cannot brute force an attack by starting multiple requests in parallel?

AFAIK the brute force protection would catch those parallel requests, and increase the failure counters despite those being concurrent. Having a global lock would be quite expensive, but I could see that in a separate issue we could have a lock per Keycloak node to avoid process requests for the same user only sequentially with a single Keycloak node.